Fix for skipped test display.
It's not quite ideal yet. The ResultSet is heirarhical but I'm not sure if
a single nested test should mark an entire suite as 'skipped'.

Improved documentation for the Context class.

Added the Belt library, mainly for Belt\Trace which is usefulf or debugging.

Changed the status codes to use letters, which are slightly easier to grok than
numbers when you're looking at the raw result data.